requirements.txt
discord.py
google-api-python-client
google-cloud-vision
google-cloud-translate
opencv-python
googlemaps